+++ init.php
@@ -6,10 +6,12 @@ require(dirname(__FILE__) . '/lib/Resource.php');
 
 // Resources
 require(dirname(__FILE__) . '/lib/Authorizations.php');
+require(dirname(__FILE__) . '/lib/Analytics.php');
 require(dirname(__FILE__) . '/lib/Collections.php');
 require(dirname(__FILE__) . '/lib/Customers.php');
 require(dirname(__FILE__) . '/lib/Products.php');
 require(dirname(__FILE__) . '/lib/Videos.php');
+require(dirname(__FILE__) . '/lib/Watchlist.php');
 
 // Errors
 require(dirname(__FILE__) . '/lib/Error/Base.php');